Add generator for store contexts/providers
