+++ context/app/static/js/helpers/zustand/index.ts
@@ -1,3 +1,3 @@
 export * from './middleware';
 export * from './types';
-export * from './use-context-store';
+export * from './create-context-store';
